# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Documentation for the InterMine Registry RESTful API.",
    "version": "1.0.0",
    "title": "Intermine Registry",
    "license": {
      "name": "GNU General Public License v3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    }
  },
  "host": "registry.intermine.org",
  "basePath": "/service",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/synchronize": {
        "put": {
          "tags": [
            "synchronize"
          ],
          "summary": "Update the branding & version information of every instance on the registry",
          "description": "Update the branding & version information of every instance on the registry. For each instance on the registry we do 4 parallel http requests to the branding & version endpoints to verify if they have been updated.",
          "operationId": "synchronizeInstances",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": "Instance Version & Branding Successfully Updated on the Registry"
            },
            "406": {
              "description": "Bad Request"
            }
          },
          "security": [
          ]
        }
      },
      "/synchronize/{instanceIdOrName}": {
          "put": {
            "tags": [
              "synchronize"
            ],
            "summary": "Update the branding & version information of the specified instance on the registry",
            "description": "Update the branding & version information of the instance given in the URL",
            "operationId": "synchronizeInstance",
            "consumes": [
              "application/json"
            ],
            "parameters": [
              {
                "name": "instanceIdOrName",
                "in": "path",
                "description": "ID or Name of instance to update",
                "required": true,
                "type": "string"
              }
            ],
            "produces": [
              "application/json"
            ],
            "responses": {
              "201": {
                "description": "Instances Versions & Branding Successfully Updated on the Registry"
              },
              "404": {
                "description": "Instance Not Found"
              },
              "406": {
                "description": "Bad Request"
              }
            },
            "security": [
            ]
          }
        }
  },
  "securityDefinitions": {
  },
...
}